Add file scopes for the Android plugin
This CL adds a couple of scopes to the project which can be used
to for example do file coloring. For example, the obsoleteplugin
scope can be used to highlight files in the old version of the
Android plugin as orange, such that you don't accidentally edit
it. Similarly, the platform scope highlights non-Android plugin
files, so you can be aware that you're now editing something
outside the plugin.

Change-Id: I5a29bc494a7d22b15aaa8900c65706728f7ece0c