Add support for deprecated palette items.
PaletteItems can now be marked as deprecated for a certain version,
along with an optional deprecation hint. This can then be used
by designer panel implementations to decide whether the item
is deprecated for the current version, and if so, the corresopnding
palette items are shown as strikethrough in the palette -- and the
tooltip will show the deprecation hint.

Change-Id: Icdc46f5af9aa2e6cc17f9b7aa1e62fb3c809e4a4