Add parent deletion handler
This changeset adds a new interface, IComponentDeletionParticipant.
Layout RadComponents can implement this interface to be able to
participate when one or more of their children are being deleted.

For example, in RelativeLayout, if you remove a child, then any
*other* references to the deleted child should be updated.

This is similar to the existing IGroupDeleteComponent interface,
but that interface applies to the type of components being deleted,
whereas the new interface applies to the type of the parent. Since
you can select many components across different layouts, the delete
handler code will segment the selection into individual sets of
components under each parent, then call each parent in turn with
the set of relevant deleted children.

Change-Id: If6e91b75121e3c1839e0506a7fd5e9972f82a80e