Add modifier keys to operations
This changeset records the modifier key mask, and makes it available
to InputTools as well as operations (such as drop and resize
operations).  This can be used during a resize operation to for
example implement traditional sizing modifiers (e.g. shift to
constrain shape, alt to center around the starting point). The marqee
selection also implements centering.

Change-Id: I15d640fa584d786d071d3f6fe25565f745a0d14a