Add component grouping utility, and use it for deletion
This partitions the selection into sets of siblings for each parent
(and since this is already done for deletion, reuse the utility there
and add a unit test). It also makes the selection action construction
overrideable; in the Android designer we'll pass siblings rather than
the whole selection to each container.

Change-Id: I51b4fd10026b01c7d63b3bc9065dff2170e003db